requirements.txt
Django~=4.2.16
requests~=2.32.3
djangorestframework~=3.15.2